requirements.txt
# Core data processing and analysis
pandas>=2.3.1
numpy>=2.3.2

# Visualization
matplotlib>=3.10.5
seaborn>=0.13.2

# Audio processing
librosa>=0.11.0

# Scientific computing
scipy>=1.16.1

# Machine learning
torch>=2.6.0
torchaudio>=2.6.0
scikit-learn>=1.7.1

# Data formats
pyarrow>=18.0.0

PyYAML>=6.0.2
pudb>=2025.1
# Note: argparse, os, pathlib, and collections are built-in Python modules